requirements.txt
langchain_google_genai
langchain
langchain_community
python-dotenv
streamlit